strash: Make it work with newer `find` versions
The behaviour of `@` in `%T@` and friends has changed: it can now return a
fractional part (and usually does), which is undesired for this script since
it uses the shell's `test -eq` functionality for comparison, which expects
integers.

Given that 'libtrash' is for GNU/Linux, we can expect that recent `strftime`
implementations are available, providing the `s` modifier to return integer
values.